refactor: 의존성 줄이는 방향으로 패키지 구조 변경
