+++ src/main/java/lotto/controller/dto/LottoWinningNumberCreateRequest.java
@@ -1,4 +1,4 @@
-package lotto.view.dto;
+package lotto.controller.dto;
 
 import java.util.ArrayList;
 import java.util.List;
