+++ src/main/java/lotto/view/LottoGameConsoleView.java
@@ -3,12 +3,12 @@ package lotto.view;
 import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
 import lotto.controller.LottoGameView;
+import lotto.controller.dto.LottoBonusNumberCreateRequest;
+import lotto.controller.dto.LottoPurchaseRequest;
+import lotto.controller.dto.LottoResponse;
+import lotto.controller.dto.LottoWinningNumberCreateRequest;
+import lotto.controller.dto.LottoWinningStatistics;
 import lotto.domain.LottoPrize;
-import lotto.view.dto.LottoBonusNumberCreateRequest;
-import lotto.view.dto.LottoPurchaseRequest;
-import lotto.view.dto.LottoResponse;
-import lotto.view.dto.LottoWinningNumberCreateRequest;
-import lotto.view.dto.LottoWinningStatistics;
 
 public class LottoGameConsoleView implements LottoGameView {
